NoteShow renamed to ShowNote
