+++ src/components/views/notes/ShowNote.js
